Implement WebSocket response
The response significantly contains a handler which is passed to the
server to be run after upgrading the connection. The handler runs
concurrently and must be written in a (tail-)recursive style to keep
running after processing each message.

All made possible thanks to Antonio Monteiro's amazing work on
websocketaf et al.