Re-implement topics with ephemeral caches
This greatly simplifies both the interface and implementation, as users
no longer need to remember to unsubscribe from streams and close topics.
Partly also because we no longer expose a stream to iterate over, but
instead just a `pull` function to get the next message in the topic with
a timeout, just like in WebSocket handlers.