Pull incoming message from WebSocket instead of simultaneous timeout
When a client sends a message to ReWeb's WebSocket connection and the
message `pull` simultaneously times out, return the message instead of
timing out.

We need to use `Lwt.npick` instead of `Lwt.nchoose` because if the
timeout expires, we still want to _cancel_ the promise that's waiting to
get the message so it doesn't keep waiting and then removes the message
from the stream; we want to leave the message for the handler to process
next time it can.

Fix #32